feat(messages): implement message handling routes and context integration
